Extract the node server script to a dedicated bin
Instead of configuring the host, port and node paths by performing
string replacements in the JS source code, the script now receives the
host and port through environment variables, and the node paths are
handled through the builtin NODE_PATH environment variable.

This change opens the way to being able to run the server manually
separately (as it is now a working node script which can be launched
separately), but also to a bigger refactoring of the driver to move more
logic to the node script instead of sending it code to eval.
Such refactoring will come later.
